requirements.txt
aiogram
httpx
emoji-country-flag
ipwhois